Inroduce Websocket Client
Introduce a basic websocket client able to connect to a remote peer, send and receive payloads and disconnect

Each websocket client is created after supplying a WebsocketMsgHandler and a WebsocketPayloadRegistry.

Registry is used in order to specify which payload the websocket client will handle. MsgHandler function will be executed once a valid payload is received.

For a more smooth disconnection, internal messages also exist. Upon disconnection, client wiill send a goodbye to other peer and other peer will acknowledge it. A slot is provided through MsgHandler for clean-up actions

For performance reasons, reading and sending of messages are done through pumping in two separate goroutines. Handling of messages is done also in a separate goroutine to not block reading of next messages

Clean-Up and disconnection notification is handled through Once in order to be sure they execute only once upon the first disconnection. This is necessary since read and send is handle in differnt goroutines

A traffic controller decides if messages are allowed to be sent. In case that client disconnects, only Goodbye and GoodbyeAck messages will be allowed to be sent or received. No message will be processed if client has not yet been connected